Registering handlers
To register a handler we use the `.on` method on our emitter. The first argument tells the emitter what event this handler is for. When an event is triggered (`.emit`ted) the first argument selects what handlers to execute and the rest is passed as arguments to the first handler.

You should be able to see something like this on the console:

```
"This is a log message [" Arguments { 0: "string", 1: 123, 2: false, 2 more… } "]" index.js:10:6
"This is another log message [" Arguments { 0: "gnirts", 1: 321, 2: true, 2 more… } "]" index.js:14:6
```